*: factor out separate 'desktop' wInd3x command
In preparation for a wasm build, we:

 1. Move all CLI code to cmd/wInd3x
 2. Factor out a common USB interface that's implemented by the CLI code
    to be backed by gousb
 3. Clean up some of the 'TODO's around application naming

With time, we will also start refactoring most of the logic from cmd_*
files into a 'unified' command system that can be triggered from both
CLI and wasm builds. But for that we have to first settle on some kind
of USB device lifecycle management system, which is drastically
different on CLI and wasm, and I'm not ready to design that before we
actually have _some_ wasm port.